refactor(titlebar): extract tooltip coordinator and update-pill state into composables
Pure relocation. TitleBarApp.vue 1090 -> 900 lines.

- useTitleBarTooltip: macOS hover tooltip dispatcher (issue #514) + tooltipAttrs helper
- useUpdatePills: app-update + install-update pill state, labels, click handlers, and bridge subscriptions

Amp-Thread-ID: https://ampcode.com/threads/T-019e2311-6aeb-7349-9fd9-560288cfbe38
