+++ src/renderer/src/comfyTitleBar/TitleBarApp.vue
@@ -10,6 +10,8 @@ import {
   RefreshCw,
 } from 'lucide-vue-next'
 import { installTypeMetaFor } from '../lib/installTypeIcon'
+import { useTitleBarTooltip } from './useTitleBarTooltip'
+import { useUpdatePills } from './useUpdatePills'
 
 const { t } = useI18n()
 
@@ -247,78 +249,16 @@ const showInstallTypeIcon = computed(
 const themeBg = ref<string | null>(null)
 const themeText = ref<string | null>(null)
 
-/**
- * Title-bar status pills.
- *
- * The app-update pill (right of the hamburger) shows when the
- * auto-updater has either downloaded an update (`'ready'`, prompts
- * Restart-to-update via the popover) or detected one is available
- * (`'available'`, prompts Download via the popover). State is pushed
- * from main on `comfy-titlebar:app-update-state-changed`; the pill
- * disappears entirely when `kind` is `null` so the title bar reads
- * clean in the steady state.
- *
- * The install-update pill (right of the install pill in the center)
- * fires when the active install's `statusTag.style === 'update'` —
- * the same signal the chooser tile's "Update" pill consumes. State is
- * pushed from main on `comfy-titlebar:install-update-changed` and is
- * gated on `!isInstallLess` (install-less hosts have no install backing
- * the window, so an install-scoped pill is meaningless there).
- */
-const appUpdateState = ref<{
-  kind: 'available' | 'downloading' | 'ready' | null
-  version: string | null
-  autoUpdate: boolean
-}>({ kind: null, version: null, autoUpdate: true })
-const installUpdateState = ref<{ available: boolean; version: string | null }>({
-  available: false,
-  version: null,
-})
-
-const appUpdatePillLabel = computed<string | null>(() => {
-  const s = appUpdateState.value
-  if (!s.kind) return null
-  if (s.kind === 'ready') return t('titleBar.desktopUpdateReady')
-  if (s.kind === 'downloading') return t('titleBar.desktopUpdateDownloading')
-  // 'available' — only fires with auto-updates OFF (main suppresses
-  // it when ON and triggers the download itself).
-  return t('titleBar.desktopUpdateAvailable')
-})
-
-/** Tooltip / aria-label augments the pill label with the version when
- *  one is known, so the compact pill stays scan-friendly while the
- *  full "Desktop Update Ready (v1.2.3)" detail is one hover away. */
-const appUpdatePillTooltip = computed<string>(() => {
-  const label = appUpdatePillLabel.value
-  if (!label) return ''
-  const v = appUpdateState.value.version
-  return v
-    ? t('titleBar.desktopUpdateWithVersion', { label, version: v })
-    : label
-})
-
-/** Install-update pill copy. Mirrors the app-update pill's
- *  "Update {version}" format when main carries a target version
- *  through the install's status tag, falling back to the generic
- *  "Update available" label when no version is known. */
-const installUpdatePillLabel = computed<string>(() => {
-  const v = installUpdateState.value.version
-  return v
-    ? t('titleBar.installUpdateVersion', { version: v })
-    : t('titleBar.installUpdateAvailable')
-})
-
-const showAppUpdatePill = computed(() => appUpdateState.value.kind !== null)
-const showInstallUpdatePill = computed(
-  () => !isInstallLess.value && installUpdateState.value.available,
-)
-
-function handleAppUpdatePill(): void {
-  bridge?.clickAppUpdatePill()
-}
-function handleInstallUpdatePill(): void {
-  bridge?.clickInstallUpdatePill()
-}
+const {
+  appUpdateState,
+  appUpdatePillLabel,
+  appUpdatePillTooltip,
+  installUpdatePillLabel,
+  showAppUpdatePill,
+  showInstallUpdatePill,
+  handleAppUpdatePill,
+  handleInstallUpdatePill,
+} = useUpdatePills({ bridge, isInstallLess })
 
 /**
  * Title-bar downloads tray. Always-visible icon button sitting in the
@@ -420,141 +360,10 @@ function anchorBelow(el: HTMLElement | null | undefined): MenuAnchor {
   return { x: Math.round(rect.left), y: Math.round(rect.bottom) }
 }
 
-/**
- * Issue #514 — macOS hover-tooltip plumbing.
- *
- * On macOS the native HTML `title` tooltip does not reliably fire for
- * controls inside a sibling chrome `WebContentsView` that isn't the
- * focused web contents (Electron + Cocoa quirk). The title bar always
- * sits in such a view, so on macOS we route hover through main, which
- * positions a cached `WebContentsView` popup attached to the host
- * window — that popup escapes the title-bar view's 37px clip. On
- * Windows / Linux the native `title` attribute renders Chromium's own
- * tooltip widget reliably; the JS handlers below are no-ops in that
- * case so we don't end up with two tooltips.
- *
- * Implementation is delegated: a single `pointermove` / `pointerleave`
- * pair on the header root finds the closest `[data-title-tooltip]`
- * ancestor and fires `showTip` / `hideTip`. New tooltipped elements
- * just need the data attribute — no per-element wiring.
- */
-/** Initial show delay (ms). Matches the cadence of native HTML
- *  tooltips on macOS / Win so a quick fly-by across the title bar
- *  doesn't flash bubbles. */
-const TOOLTIP_SHOW_DELAY_MS = 400
-/** Hover-handoff window (ms). If a tooltip was visible up to this
- *  long ago, the next hover over a different tooltipped element shows
- *  immediately — same convention as native macOS / browser tooltips,
- *  where the first hover earns the wait but subsequent ones in a
- *  scanning gesture feel snappy. */
-const TOOLTIP_HANDOFF_WINDOW_MS = 1500
-let tooltipShowTimer: number | null = null
-/** Text of the tooltip the renderer most recently asked main to show
- *  (or queue). `null` while nothing is pending or visible. */
-let activeTooltipText: string | null = null
-/** True between `bridge.showTooltip()` and the corresponding
- *  `bridge.hideTooltip()` — i.e., a tooltip is currently visible. The
- *  pending-but-not-yet-shown state has this `false`. */
-let isTooltipVisible = false
-/** `performance.now()` timestamp of the most recent
- *  `bridge.hideTooltip()`. Drives the hover-handoff fast path. */
-let lastHiddenAt = -Infinity
-
-/**
- * Single source of truth for tooltip-related attributes on title-bar
- * controls. Cocoa's native HTML `title` tooltip occasionally DOES
- * fire for our sibling-view buttons even though it's documented as
- * unreliable; when it does, the user gets two bubbles at once (the
- * native one plus our custom popup). Keep them mutually exclusive at
- * the source by emitting `title` only off-mac and `data-title-tooltip`
- * only on mac. `aria-label` is unconditional so screen readers see
- * the same string regardless of platform — pass `ariaLabel` separately
- * when the visible label and the tooltip copy intentionally differ
- * (e.g. the Send Feedback button shows "Beta Feedback" but the
- * tooltip reads "Send Feedback").
- */
-function tooltipAttrs(text: string, ariaLabel?: string): Record<string, string> {
-  const base: Record<string, string> = { 'aria-label': ariaLabel ?? text }
-  if (isMac.value) {
-    base['data-title-tooltip'] = text
-  } else {
-    base.title = text
-  }
-  return base
-}
-
-function findTooltipTarget(target: EventTarget | null): {
-  text: string
-  rect: DOMRect
-} | null {
-  if (!(target instanceof Element)) return null
-  const el = target.closest('[data-title-tooltip]') as HTMLElement | SVGElement | null
-  if (!el) return null
-  const text = el.getAttribute('data-title-tooltip')
-  if (!text) return null
-  return { text, rect: el.getBoundingClientRect() }
-}
-
-function cancelPendingTooltipShow(): void {
-  if (tooltipShowTimer !== null) {
-    window.clearTimeout(tooltipShowTimer)
-    tooltipShowTimer = null
-  }
-}
-
-function hideTip(): void {
-  cancelPendingTooltipShow()
-  if (activeTooltipText === null) return
-  activeTooltipText = null
-  if (isTooltipVisible) {
-    isTooltipVisible = false
-    lastHiddenAt = performance.now()
-  }
-  bridge?.hideTooltip()
-}
-
-function fireShowTooltip(text: string, rect: DOMRect): void {
-  bridge?.showTooltip({
-    text,
-    leftX: Math.round(rect.left),
-    rightX: Math.round(rect.right),
-    bottomY: Math.round(rect.bottom),
-  })
-  isTooltipVisible = true
-}
-
-function handleTooltipPointer(event: PointerEvent): void {
-  if (!isMac.value) return
-  const found = findTooltipTarget(event.target)
-  if (!found) {
-    hideTip()
-    return
-  }
-  if (found.text === activeTooltipText) {
-    // Same trigger as before — no work needed. (Either we're still
-    // waiting on the show timer, or the tooltip is already visible;
-    // either way we don't reset state mid-hover.)
-    return
-  }
-  // Different (or first) tooltipped target. Hide any in-flight tooltip
-  // and queue the new one. If we were just showing a tooltip moments
-  // ago (hover-handoff), skip the show delay so scanning across the
-  // title bar feels instant — matches native macOS behaviour.
-  const handoff =
-    isTooltipVisible || performance.now() - lastHiddenAt < TOOLTIP_HANDOFF_WINDOW_MS
-  hideTip()
-  const captured = found
-  activeTooltipText = captured.text
-  if (handoff) {
-    fireShowTooltip(captured.text, captured.rect)
-    return
-  }
-  tooltipShowTimer = window.setTimeout(() => {
-    tooltipShowTimer = null
-    if (activeTooltipText !== captured.text) return
-    fireShowTooltip(captured.text, captured.rect)
-  }, TOOLTIP_SHOW_DELAY_MS)
-}
+const { tooltipAttrs, handleTooltipPointer, hideTip } = useTitleBarTooltip({
+  bridge,
+  isMac,
+})
 
 function handleFileMenu(): void {
   // Hide any in-flight tooltip — the menu will obscure the same area
@@ -585,8 +394,6 @@ let unsubFullscreen: (() => void) | undefined
 let unsubMenuOpened: (() => void) | undefined
 let unsubMenuClosed: (() => void) | undefined
 let unsubFirstUseMode: (() => void) | undefined
-let unsubAppUpdate: (() => void) | undefined
-let unsubInstallUpdate: (() => void) | undefined
 let unsubDownloads: (() => void) | undefined
 
 /** Drop the hover gate immediately when input leaves the title-bar
@@ -643,12 +450,6 @@ onMounted(() => {
   unsubFirstUseMode = bridge.onFirstUseModeChanged((mode) => {
     firstUseMode.value = mode
   })
-  unsubAppUpdate = bridge.onAppUpdateStateChanged((next) => {
-    appUpdateState.value = next
-  })
-  unsubInstallUpdate = bridge.onInstallUpdateAvailable((next) => {
-    installUpdateState.value = next
-  })
   unsubDownloads = bridge.onDownloadsChanged((next) => {
     downloadsState.value = next
   })
@@ -671,8 +472,6 @@ onUnmounted(() => {
   unsubMenuOpened?.()
   unsubMenuClosed?.()
   unsubFirstUseMode?.()
-  unsubAppUpdate?.()
-  unsubInstallUpdate?.()
   unsubDownloads?.()
   window.removeEventListener('blur', handleWindowBlur)
   window.removeEventListener('pointermove', handlePointerMove)
